bot/requirements.txt
delta-rest-client==1.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
colorama==0.4.6
cryptography==41.0.7